feat(overhaul): market-ready redesign — 66 files, 6-agent parallel implementation
WIZARD (7-Turn Complete):
- Fix Turn 4: brand_data now transmitted → sector_affinity functional
- Add Steps 5-7: investment amount, ESG preference, income preference
- Router constraint raised le=4 → le=7 with handlers for all turns
- Confidence score: base 0.20 when Turn 3 done, 0.60 when all 7 done
- Reveal screen shows all 7 profile dimensions

SIGNALS:
- BUY/HOLD/WATCH → BUY/HOLD/SELL throughout (user-facing labels)
- Decision: default sort by PRISMA score descending
- Decision: BUY/HOLD/SELL tooltips + SHAP tooltip + Confidence tooltip
- Empty state when no signals with CTA to /rankings

BACKTEST:
- Fix NaN charts: list[Decimal] → list[float] in schema
- Frontend: remove now-redundant parseFloat() calls
- Tooltips: CAGR, Volatilität, Sharpe Ratio, Max. Drawdown, Alpha, Benchmark

DASHBOARD:
- Quick-start cards: Ranking / Signale / Portfolio
- Top 3 BUY signals section with score bars + links
- MacroWidget: SNB rate context hint (Tiefzins / Zinsen gestiegen)
- StatsCards: info tooltips on all metrics

DISCOVER:
- Empty state with CTA when no wizard profile
- Stock cards: signal badge + score bar + sector tag + 3a badge
- Info tooltip on page title

PORTFOLIO:
- Real-time weight indicator: green (100%) / yellow (<100%) / red (>100%)
- Submit disabled when totalWeight > 100 or = 0
- AlertTriangle icon instead of ⚠️ emoji
- Empty state when no positions

FACTSHEET + MEMO:
- Score interpretation labels (Ausgezeichnet/Gut/Durchschnittlich/Schwach)
- AuditPanel: tooltips on all 4 scores
- MemoEmpty: improved CTA with Sparkles icon
- MemoContent: "Generiert am" timestamp
- MemoPanel: renamed to "KI-Analyse" with disclaimer tooltip

INFO TOOLTIPS (Otto-Normal-User):
- Rankings: Universe, Run, Composite Score, Signal, Value/Quality/Income Score
- Alerts: Ticker, Threshold, Trigger
- Fonds: TER, Sharpe Ratio, ETF explanation
- News: RAG explanation, search hint
- Research: similarity score, memo explanation
- Steuer: Verrechnungssteuer, Haltedauer, Steuerprofil
- Portfolio: Gewichtung, Beta, Rebalancing

DESIGN:
- All emojis removed → Lucide fine-line icons (Compass, AlertTriangle, Globe)
- Rank medals 🥇🥈🥉 → numbered circle badges
- 🧭 in nav/wizard → <Compass /> icon
- Swiss flag 🇨🇭 → "CH" text + Globe icon
- Chat tool hints: emojis removed
- SBB title typography: trailing period on all 18 page h1 titles
  (Dashboard. / Ranking starten. / Decision Intelligence. / etc.)

+++ backend/interfaces/rest/schemas/backtest.py
@@ -1,7 +1,6 @@
 """Pydantic-Schemas für Backtest-REST-Endpunkte."""
 
 from datetime import date
-from decimal import Decimal
 from typing import Literal
 from uuid import UUID
 
@@ -20,11 +19,11 @@ class RunBacktestRequest(BaseModel):
 
 
 class PortfolioMetricsResponse(BaseModel):
-    total_return: Decimal
-    cagr: Decimal
-    annual_vol: Decimal
-    sharpe: Decimal
-    max_drawdown: Decimal
+    total_return: float
+    cagr: float
+    annual_vol: float
+    sharpe: float
+    max_drawdown: float
 
     @classmethod
     def from_entity(cls, m: PortfolioMetrics) -> "PortfolioMetricsResponse":
@@ -33,9 +32,9 @@ class PortfolioMetricsResponse(BaseModel):
 
 class BacktestSeriesResponse(BaseModel):
     dates: list[date]
-    prisma: list[Decimal]
-    universe: list[Decimal]
-    benchmark: list[Decimal]
+    prisma: list[float]
+    universe: list[float]
+    benchmark: list[float]
 
 
 class BacktestResultResponse(BaseModel):
